// web-client/src/app/data/error/field-names.enum.ts
export enum FieldNames {
  LICENSE_CODE = "licenseCode",
  USERNAME = "username",
  PASSWORD = "password",
  FIRST_NAME = "firstName",
  LAST_NAME = "lastName",
  CITY = "city",
  STREET_ADDRESS_1 = "streetAddress1",
  STREET_ADDRESS_2 = "streetAddress2",
  ZIP_CODE = "zipCode",
  REGION = "region",
  CONTACT_NUMBER = "contactNumber",
  PESEL = "peselNumber"
}