// web-client/src/app/data/error/field-names-pl.enum.ts
export enum FieldNamesPL {
  LICENSE_CODE = "kod licencji",
  USERNAME = "nazwa użytkownika",
  PASSWORD = "hasło użytkownika",
  FIRST_NAME = "imię",
  LAST_NAME = "nazwisko",
  CITY = "miasto",
  STREET_ADDRESS_1 = "adres",
  STREET_ADDRESS_2 = "adres2",
  ZIP_CODE = "kod pocztowy",
  REGION = "województwo",
  CONTACT_NUMBER = "numer telefonu",
  PESEL = "pesel"
}