// web-client/src/app/data/error/messages-pl.enum.ts
export enum MessagesPL {
  NOT_BLANK = "nie może być puste",
  SIZE_1_20 = "musi zawierać od 1 do 20 znaków",
  SIZE_0_60 = "musi zawierać od 0 do 60 znaków",
  SIZE_1_60 = "musi zawierać od 1 do 60 znaków",
  INVALID_PHONE = "musi składać się z cyfr (maks. 20) lub ze znaku '+' i cyfr (maks. 20)",
  USERNAME_UNIQUE = "już istnieje",
  INVALID_PESEL = "musi składać się z 11 cyfr"
}