// web-client/src/app/data/error/messages.enum.ts
export enum Messages {
  NOT_BLANK = "must not be blank",
  SIZE_1_20 = "must consist of 1 to 20 characters",
  SIZE_0_60 = "must consist of 0 to 60 characters",
  SIZE_1_60 = "must consist of 1 to 60 characters",
  INVALID_PHONE = "must consist of digits (max 20) or '+' and digits (max 20)",
  USERNAME_UNIQUE = "must be unique",
  INVALID_PESEL = "must consist of 11 digits"
}